+++ src/model/__tests__/PhoneticAlphabet.test.ts
@@ -1,0 +1,7 @@
+import phoneticAlphabet from '../PhoneticAlphabet';
+
+test('Test that the alphabet is covered', () => {
+  expect(phoneticAlphabet.map(codeWord => codeWord.letter)).toEqual(
+    'abcdefghijklmnopqrstuvwxyz'.split('')
+  );
+});